Fix quirks not working on the dual version
VID/PID were transmitted as 8 bits between B and A, when they should
be 16 bits.

This caused the quirks mechanism to not work on the dual version.

+++ firmware/src/dual.h
@@ -14,8 +14,8 @@ enum class DualCommand : uint8_t {
 
 struct __attribute__((packed)) device_connected_t {
     DualCommand command = DualCommand::DEVICE_CONNECTED;
-    uint8_t vid;
-    uint8_t pid;
+    uint16_t vid;
+    uint16_t pid;
     uint8_t dev_addr;
     uint8_t interface;
     uint8_t report_descriptor[0];
